fix(BUG-016): implement cross-platform fixes M1, M2, M6, M7
- M1: Add CUDA guards to seeding.py (torch.cuda.is_available checks)
- M2: Add dtype conversion to loss.py (class_weights match logits.dtype)
- M6: Add --hf-cache CLI option to redirect HuggingFace hub downloads
- M7: Unify disk size documentation across DATA.md, TROUBLESHOOTING.md, data/README.md

Fixes prevent:
- Spurious CUDA calls on MPS/CPU systems
- Dtype mismatch when loss is called outside autocast
- WSL2 disk-full errors from hub cache filling ~/.cache/huggingface
- Confusion over 273GB vs 50GB disk requirements

+++ src/arc_meshchop/training/loss.py
@@ -56,9 +56,11 @@ class WeightedCrossEntropyLoss(nn.Module):
         Returns:
             Scalar loss value.
         """
-        # Initialize loss function on first call (to get correct device)
+        # Initialize loss function on first call (to get correct device/dtype)
         if self._loss_fn is None or self.class_weights.device != logits.device:
-            self.class_weights = self.class_weights.to(logits.device)
+            self.class_weights = self.class_weights.to(
+                logits.device, dtype=logits.dtype
+            )
             self._loss_fn = nn.CrossEntropyLoss(
                 weight=self.class_weights,
                 label_smoothing=self.label_smoothing,
